requirements.txt
fastapi
openai
python-dotenv
black
uvicorn